HW-1 1. listener 변수 삭제, 파라미터로 추가. 인터페이스 분리
