+++ wswon/AndroidArchitectureStudy/app/src/main/java/com/architecture/study/server/UpbitRequestListener.kt
@@ -1,0 +1,7 @@
+package com.architecture.study.server
+
+interface UpbitRequestListener<T> {
+    fun onSucess(dataList: List<T>)
+    fun onEmpty(str: String)
+    fun onFailure(str: String)
+}
